+++ lib/hooks/useScene/__tests__/useScene.test.ts
@@ -688,7 +688,7 @@ describe("useScene", () => {
       players: [{ character: {} }, { character: {} }],
     });
     expect(
-      result.current.useCharacters.actions.updateIfExists
+      result.current.useCharacters.actions.addOrUpdateIfMoreRecent
     ).toHaveBeenCalledTimes(2);
   });
 
@@ -742,7 +742,7 @@ function mockUseCharacters() {
     actions: {
       add: jest.fn(),
       upsert: jest.fn(),
-      updateIfExists: jest.fn(),
+      addOrUpdateIfMoreRecent: jest.fn(),
       remove: jest.fn(),
       duplicate: jest.fn(),
       select: jest.fn(),